First commit.
This is a simple tool that reads and returns the http code from
the header of a website url request.  Additionally it provides
a count o fhte number of urls visited and the number of times it
reported on the urls.
The program reads an input file of return seperated urls from a
file called urlslist.php.